Move kernels/interpolation.py to interpolators.py
