+++ tests/test_diffusion.py
@@ -8,7 +8,7 @@ from parcels._datasets.structured.generated import simple_UV_dataset
 from parcels.field import Field, VectorField
 from parcels.fieldset import FieldSet
 from parcels.kernels import AdvectionDiffusionEM, AdvectionDiffusionM1, DiffusionUniformKh
-from parcels.kernels.interpolation import XLinear
+from parcels.interpolators import XLinear
 from parcels.particle import Particle, Variable
 from parcels.particleset import ParticleSet
 from parcels.xgrid import XGrid
