+++ tests/test_field.py
@@ -5,12 +5,13 @@ import pytest
 import uxarray as ux
 import xarray as xr
 
-from parcels import Field, UXPiecewiseConstantFace, UXPiecewiseLinearNode, VectorField
+from parcels import Field, VectorField
 from parcels._datasets.structured.generic import T as T_structured
 from parcels._datasets.structured.generic import datasets as datasets_structured
 from parcels._datasets.unstructured.generic import datasets as datasets_unstructured
 from parcels.uxgrid import UxGrid
 from parcels.xgrid import XGrid
+from parcels.interpolators import UXPiecewiseConstantFace, UXPiecewiseLinearNode
 
 
 def test_field_init_param_types():
